grocery-backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==23.0.0
whitenoise==6.9.0
